nf10-newnic: add a placeholder for non-crash in ifconfig w/ ip addr
as soon as ifconfig is up with ip addr, a small packet is sent, but it
has no enough headroom to add netfpga metadata as is. So, it should be
handled by copying the packet to another memory area. Before
implementing this, added a placeholder that returns immediately in the
event of no headroom. But, it's NOT a correct way.